added standardized CRON util
