+++ apps/sim/lib/auth/internal.ts
@@ -1,5 +1,9 @@
 import { jwtVerify, SignJWT } from 'jose'
+import { type NextRequest, NextResponse } from 'next/server'
 import { env } from '@/lib/env'
+import { createLogger } from '@/lib/logs/console-logger'
+
+const logger = createLogger('CronAuth')
 
 // Create a secret key for JWT signing
 const getJwtSecret = () => {
@@ -44,4 +48,27 @@ export async function verifyInternalToken(token: string): Promise<boolean> {
     // Token verification failed
     return false
   }
+}
+
+/**
+ * Verify CRON authentication for scheduled API endpoints
+ * Returns null if authorized, or a NextResponse with error if unauthorized
+ */
+export function verifyCronAuth(request: NextRequest, context?: string): NextResponse | null {
+  const authHeader = request.headers.get('authorization')
+  const expectedAuth = `Bearer ${env.CRON_SECRET || 'your-cron-secret'}`
+
+  if (authHeader !== expectedAuth) {
+    const contextInfo = context ? ` for ${context}` : ''
+    logger.warn(`Unauthorized CRON access attempt${contextInfo}`, {
+      providedAuth: authHeader,
+      ip: request.headers.get('x-forwarded-for') || request.headers.get('x-real-ip') || 'unknown',
+      userAgent: request.headers.get('user-agent') || 'unknown',
+      context,
+    })
+
+    return NextResponse.json({ error: 'Unauthorized' }, { status: 401 })
+  }
+
+  return null
 }
